pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-experiments"
version = "0.1.0"
description = "ML experiments framework with Modal GPU integration, WebDataset streaming, and comprehensive benchmarking"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "ML Experiments Team"}
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "modal>=0.55.0",
    "webdataset>=0.2.0",
    "boto3>=1.29.0",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
    "bitsandbytes>=0.41.0",
    "torchmetrics>=1.2.0",
    "clean-fid>=0.1.35",
    "wandb>=0.16.0",
    "pyyaml>=6.0",
    "lpips>=0.1.4",
    "clip-score>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.scripts]
ml-experiments = "scripts.run_experiment:main"
ml-benchmark = "scripts.benchmark:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ml_experiments*", "scripts*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
